+++ colorpicker-library/src/com/saggitt/colorpickerx/AlphaPatternDrawable.java
@@ -1,0 +1,117 @@
+/*
+ *  This file is part of ColorPickerX
+ *  Copyright (c) 2021   Saul Henriquez
+ *
+ *  This program is free software: you can redistribute it and/or modify
+ *  it under the terms of the GNU General Public License as
+ *  published by the Free Software Foundation, either version 3 of the
+ *  License, or (at your option) any later version.
+ *
+ *  This program is distributed in the hope that it will be useful,
+ *  but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ *  GNU General Public License for more details.
+ *
+ *  You should have received a copy of the GNU General Public License
+ *  along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
+package com.saggitt.colorpickerx;
+
+import android.graphics.Bitmap;
+import android.graphics.Bitmap.Config;
+import android.graphics.Canvas;
+import android.graphics.ColorFilter;
+import android.graphics.Paint;
+import android.graphics.Rect;
+import android.graphics.drawable.Drawable;
+
+/**
+ * This drawable will draw a simple white and gray chessboard pattern.
+ * It's the pattern you will often see as a background behind a partly transparent image in many applications.
+ */
+class AlphaPatternDrawable extends Drawable {
+
+    private int rectangleSize = 10;
+
+    private final Paint paint = new Paint();
+    private final Paint paintWhite = new Paint();
+    private final Paint paintGray = new Paint();
+
+    private int numRectanglesHorizontal;
+    private int numRectanglesVertical;
+
+    /**
+     * Bitmap in which the pattern will be cached.
+     * This is so the pattern will not have to be recreated each time draw() gets called.
+     * Because recreating the pattern i rather expensive. I will only be recreated if the size changes.
+     */
+    private Bitmap bitmap;
+
+    AlphaPatternDrawable(int rectangleSize) {
+        this.rectangleSize = rectangleSize;
+        paintWhite.setColor(0xFFFFFFFF);
+        paintGray.setColor(0xFFCBCBCB);
+    }
+
+    @Override
+    public void draw(Canvas canvas) {
+        if (bitmap != null && !bitmap.isRecycled()) {
+            canvas.drawBitmap(bitmap, null, getBounds(), paint);
+        }
+    }
+
+    @Override
+    public int getOpacity() {
+        return 0;
+    }
+
+    @Override
+    public void setAlpha(int alpha) {
+        throw new UnsupportedOperationException("Alpha is not supported by this drawable.");
+    }
+
+    @Override
+    public void setColorFilter(ColorFilter cf) {
+        throw new UnsupportedOperationException("ColorFilter is not supported by this drawable.");
+    }
+
+    @Override
+    protected void onBoundsChange(Rect bounds) {
+        super.onBoundsChange(bounds);
+        int height = bounds.height();
+        int width = bounds.width();
+        numRectanglesHorizontal = (int) Math.ceil((width / rectangleSize));
+        numRectanglesVertical = (int) Math.ceil(height / rectangleSize);
+        generatePatternBitmap();
+    }
+
+    /**
+     * This will generate a bitmap with the pattern as big as the rectangle we were allow to draw on.
+     * We do this to chache the bitmap so we don't need to recreate it each time draw() is called since it takes a few
+     * milliseconds
+     */
+    private void generatePatternBitmap() {
+        if (getBounds().width() <= 0 || getBounds().height() <= 0) {
+            return;
+        }
+
+        bitmap = Bitmap.createBitmap(getBounds().width(), getBounds().height(), Config.ARGB_8888);
+        Canvas canvas = new Canvas(bitmap);
+
+        Rect r = new Rect();
+        boolean verticalStartWhite = true;
+        for (int i = 0; i <= numRectanglesVertical; i++) {
+            boolean isWhite = verticalStartWhite;
+            for (int j = 0; j <= numRectanglesHorizontal; j++) {
+                r.top = i * rectangleSize;
+                r.left = j * rectangleSize;
+                r.bottom = r.top + rectangleSize;
+                r.right = r.left + rectangleSize;
+                canvas.drawRect(r, isWhite ? paintWhite : paintGray);
+                isWhite = !isWhite;
+            }
+            verticalStartWhite = !verticalStartWhite;
+        }
+    }
+}