+++ colorpicker-library/src/com/saggitt/colorpickerx/ColorPicker.java
@@ -38,8 +38,6 @@ import androidx.core.content.ContextCompat;
 import androidx.recyclerview.widget.GridLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
 
-import com.saggitt.colorpickerlib.R;
-
 import java.lang.ref.WeakReference;
 import java.util.ArrayList;