+++ changelog.md
@@ -1,4 +1,7 @@
 ### v1.2.0 (Build 52)
 Release Date: 03-30-2021
 
-* **Change Licence to GPLv3**
+* **add: Support custom color selector**
+* **add: Spanish language**
+* **Change: check mark to drawable image instead of text**
+* **Change: Licence to GPLv3**